+++ lib/utils/test-task-mixin.js
@@ -1,16 +1,10 @@
 'use strict';
 
 module.exports = {
-  init() {
-    // Make sure to use our version of testem, rather than ember-cli's
-    this.testem = this.testem || new (require('testem'))();
-    this._super(...arguments);
-  },
+  run(options, ...unexpectedArgs) {
+    options = options || {};
+    options.configFile = options.configFile || 'testem-electron.js';
 
-  testemOptions(/* options */) {
-    let testemOptions = this._super(...arguments);
-    testemOptions.file = 'testem-electron.js';
-
-    return testemOptions;
+    return this._super(options, ...unexpectedArgs);
   },
 };
